uint16_t version

+++ clox/chunk.h
@@ -29,14 +29,14 @@ typedef enum {
 typedef struct {
     int count;
     int capacity;
-    uint8_t* code;
+    uint16_t* code;
     int* lines;
     ValueArray constants;
 } Chunk;
 
 void initChunk(Chunk* chunk);
 void freeChunk(Chunk* chunk);
-void writeChunk(Chunk* chunk, uint8_t byte, int line);
+void writeChunk(Chunk* chunk, uint16_t byte, int line);
 int addConstant(Chunk* chunk, Value value);
 
 #endif
